chore(ui/vars/export): move exporting to separate file
